Setting up the Database

+++ Section3-Setting-up-Database/storefront/tags/models.py
@@ -1,17 +1,14 @@
 from django.db import models
-from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.contenttypes.fields import GenericForeignKey
+
 
 class Tag(models.Model):
     label = models.CharField(max_length=255)
 
 
 class TaggedItem(models.Model):
-    # What tag applied to what object
     tag = models.ForeignKey(Tag, on_delete=models.CASCADE)
-    # Type (product, video, article)
-    # ID
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveBigIntegerField()
+    object_id = models.PositiveIntegerField()
     content_object = GenericForeignKey()
